Update TFTPClient.java
Added some globals for ackReceived and retries.
Added loop to getDataType.
Moved print start statements to after we format request and ask for Data Mode
Added a temp datagramPacket for the receive file - will first send the header (outboundpacket) and then send acks after that when needed.
Changed sendAck to formatAck and returns a datagramPacket now for use with the receiving file stuff